[FIXED JENKINS-15966] if a build cannot be scheduled show the reason why not a null pointer/method error.
The Exception was missing the constructor that was called in the case where we failed to schedule a job.
Added the missing constructor to the exception and added a test case.

The failure reason highlights the need for some higher managment for flows to handle blocking on downstream jobs
and not exhausting the exectutors.  This will be addressed in a extension plugin.

Change-Id: I7aaca30bc86501a44c1e0e126525be54db4a372a