Form checking for the DSL should use HTTP POST.
As the DSL can get rather large the queryParameter will eventually exceed
the maximum supported GET URL for various servers.

To defend around this use the POST method for checking the DSL.

Jenkins should automatically use post for all textareas, or if @RequirePOST is
rewuired.
set on the doCheckXXX method - however this is not the case and an immediate fix is