fix potential ConcurrentModificationException failure on task abort
It was possible for for the killRunningJobs method to run into a
ConcurrentModificationException when a flow has parallel blocks due to new
build blocks starting, and thus adding entries the jobsGraph as killRunningJobs
is attempting to abort the flow and all of its downstream jobs.

This error was always possible, if a flow with parallel blocks was aborted at
just the right moment, but the limitel concurrency feature makes it slightly
more likely.