correctly abort flows involving an "ignored" parallel block
When the result of a parallel block is ignored, all Exceptions throw by the
parallel monitor thread are ignored.  This means that when the build flow is
aborted, the InterruptedException which would normally abort the build is
suppressed when it should really be re-thrown.

The fix is to always throw Exceptions encountered during an ignore block unless
it is a JobExecutionFailureException.